Scheme independent re-write rule
Updated redirect annotation to support both http and https in rewrite rules. 
This rewrite rule is required because when WAS sends redirect (HTTP 302) responses, it always gives WAS port as in HTTP protocol there is no request port header similar to a host header and WAS tries to follow the behavior described in http://www-01.ibm.com/support/docview.wss?uid=swg21470923. But since we need the browser to be redirected without the port number as ingress controller listens only to http and https, this rewrite rule ensures that both http and https redirects are re-written without any port numbers so that they goto the default ports.